xserver xkeyboard-config libX11 fontconfig mesa plink git update Dec 15 2020
xserver          commit 65673b02ef28884da3267fe1be579c20e2cddfa6
xkeyboard-config commit bb10be1ece3881ad9d50c9b877d6b2677d7476c5
libX11           commit 78027fdb7aa07621a3b7cdf37050bae39012f2fb
fontconfig       commit 5cd11d19dfb2d901e1f6b690ae504d3bf5f5ff69
mesa             commit c314893988d4b7408383d5c0357319082c347fc6
putty            commit 353db3132f4f680ac5df742cfe9f6be80652397e

+++ mesalib/src/gallium/frontends/clover/llvm/metadata.hpp
@@ -164,7 +164,7 @@ namespace clover {
       /// Extract the int metadata node \p name corresponding to the kernel
       /// argument given by \p arg.
       ///
-      inline uint
+      inline uint64_t
       get_uint_argument_metadata(const ::llvm::Function &f,
                             const ::llvm::Argument &arg,
                             const std::string &name) {
